Misc code refactoring to follow C# conventions
