+++ MicroRuleEngine.Core.Tests/ExampleUsage.cs
@@ -24,7 +24,7 @@ namespace MicroRuleEngine.Tests
                 Operator = ExpressionType.Equal.ToString("g"),
                 TargetValue = "AUS"
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var compiledRule = engine.CompileRule<Order>(rule);
             bool passes = compiledRule(order);
             Assert.IsFalse(passes);
@@ -36,7 +36,7 @@ namespace MicroRuleEngine.Tests
             Order order = GetOrder();
 
             var type = order.GetType();
-            var members = MRE.Member.GetFields(type);
+            var members = Mre.Member.GetFields(type);
             Assert.IsTrue(
                 members.Where(x=> x.Name == "Customer.Country.CountryCode" && x.PossibleOperators.Any(y=> y.Name == "StartsWith")).Any()
             );
@@ -53,7 +53,7 @@ namespace MicroRuleEngine.Tests
                 TargetValue = "243",
                 Inputs = new List<object>() { "243" }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var compiledRule = engine.CompileRule<Order>(rule);
             bool passes = compiledRule(order);
             Assert.IsTrue(passes);
@@ -73,7 +73,7 @@ namespace MicroRuleEngine.Tests
                 Operator = ExpressionType.Equal.ToString("g"),
                 TargetValue = "AUS"
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var compiledRule = engine.CompileRule<Order>(rule);
             bool passes = compiledRule(order);
             Assert.IsTrue(passes);
@@ -105,7 +105,7 @@ namespace MicroRuleEngine.Tests
                     }
                 }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var fakeName = engine.CompileRule<Order>(rule);
             bool passes = fakeName(order);
             Assert.IsTrue(passes);
@@ -124,7 +124,7 @@ namespace MicroRuleEngine.Tests
                 Operator = "HasItem",//The Order Object Contains a method named 'HasItem' that returns true/false
                 Inputs = new List<object> { "Test" }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
 
             var boolMethod = engine.CompileRule<Order>(rule);
             bool passes = boolMethod(order);
@@ -145,7 +145,7 @@ namespace MicroRuleEngine.Tests
                 Operator = "HasItem",//The Order Object Contains a method named 'HasItem' that returns true/false
                 Inputs = new List<object> { "Test" }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
 
             var boolMethod = engine.CompileRule(typeof(Order), rule);
             bool passes =(bool)  boolMethod.DynamicInvoke(order);
@@ -176,7 +176,7 @@ namespace MicroRuleEngine.Tests
                      }
                  }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var boolMethod = engine.CompileRule<Order>(rule);
             bool passes = boolMethod(order);
             Assert.IsTrue(passes);
@@ -199,7 +199,7 @@ namespace MicroRuleEngine.Tests
                 Operator = "EndsWith",//Regular method that exists on string.. As a note expression methods are not available
                 Inputs = new List<object> { "ohn" }
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var childPropCheck = engine.CompileRule<Order>(rule);
             bool passes = childPropCheck(order);
             Assert.IsTrue(passes);
@@ -220,7 +220,7 @@ namespace MicroRuleEngine.Tests
                  Operator = "EndsWith", //Regular method that exists on string.. As a note expression methods are not available
                  Inputs = new List<object> { "ohn" }
              };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var childPropCheck = engine.CompileRule<Order>(rule);
              bool passes = childPropCheck(order);
              Assert.IsFalse(passes);
@@ -235,7 +235,7 @@ namespace MicroRuleEngine.Tests
                 Operator = "IsMatch",
                 TargetValue = @"^[a-zA-Z0-9]*$"
             };
-            MRE engine = new MRE();
+            Mre engine = new Mre();
             var regexCheck = engine.CompileRule<Order>(rule);
             bool passes = regexCheck(order);
             Assert.IsTrue(passes);
