+++ MicroRuleEngine/Placeholder.cs
@@ -1,0 +1,10 @@
+﻿namespace MicroRuleEngine
+{
+    internal static class Placeholder
+    {
+        public static int Int;
+        public static float Float;
+        public static double Double;
+        public static decimal Decimal;
+    }
+}
