+++ MicroRuleEngine/RuleValueString.cs
@@ -1,0 +1,6 @@
+﻿namespace MicroRuleEngine
+{
+    public class RuleValueString : RuleValue<string>
+    {
+    }
+}
